Rename unavailableError => unavailabilityError
