+++ postgresql.go
@@ -73,7 +73,7 @@ func (r *postgresqlResource) Await(ctx context.Context) error {
 	defer db.Close()
 
 	if err := db.Ping(); err != nil {
-		return &unavailableError{err}
+		return &unavailabilityError{err}
 	}
 
 	if val, ok := tags["tables"]; ok {
@@ -98,7 +98,7 @@ func awaitPostgreSQLTables(db *sql.DB, dbName string, tables []string) error {
 		}
 
 		if tableCnt == 0 {
-			return &unavailableError{errors.New("no tables found")}
+			return &unavailabilityError{errors.New("no tables found")}
 		}
 
 		return nil
@@ -135,7 +135,7 @@ func awaitPostgreSQLTables(db *sql.DB, dbName string, tables []string) error {
 
 	for _, t := range tables {
 		if !contains(actualTables, t) {
-			return &unavailableError{fmt.Errorf("table not found: %s", t)}
+			return &unavailabilityError{fmt.Errorf("table not found: %s", t)}
 		}
 	}
