+++ resource.go
@@ -17,41 +17,23 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-
 package main
 
 import (
-	"fmt"
-	"net/url"
-	"os/exec"
-	"strings"
-
 	"context"
+	"fmt"
 )
 
-type commandResource struct {
-	url.URL
+type resource interface {
+	fmt.Stringer
+	Await(context.Context) error
 }
 
-func (r *commandResource) Await(ctx context.Context) error {
-	cmdString, err := url.QueryUnescape(r.URL.Path)
-	if err != nil {
-		return err
-	}
-	// TODO(uwe): Splitting by space is brittle
-	cmdParts := strings.SplitN(cmdString, " ", 2)
-	if len(cmdParts) == 0 {
-		return fmt.Errorf("empty command")
-	}
-	cmd := cmdParts[0]
-	args := cmdParts[1:]
+type unavailabilityError struct {
+	Reason error
+}
 
-	if err := exec.CommandContext(ctx, cmd, args...).Run(); err != nil {
-		if exitErr, ok := err.(*exec.ExitError); ok {
-			return &unavailabilityError{exitErr}
-		}
-		return err
-	}
-
-	return nil
+// Error implements the error interface.
+func (e *unavailabilityError) Error() string {
+	return e.Reason.Error()
 }
