+++ tcp.go
@@ -35,7 +35,7 @@ func (r *tcpResource) Await(ctx context.Context) error {
 	dialer := &net.Dialer{}
 	_, err := dialer.DialContext(ctx, r.URL.Scheme, r.URL.Host)
 	if err != nil {
-		return &unavailableError{err}
+		return &unavailabilityError{err}
 	}
 
 	return nil
